+++ android/src/main/kotlin/com/svga/plugin/svga_plugin/flutter/FlutterParseCompletion.kt
@@ -19,7 +19,7 @@ class FlutterParseCompletion(
     private val registry = WeakReference(dataSource.registry)
 
     interface DataSource {
-        val widgetIdList: List<Int>
+        val widgetIdList: List<Long>
 
         val registry: TextureRegistry
